kernel/fs: changed permission system.
Previously, we passed the pid, uid and gid to userspace filesystems
on all operations and they used the isingroup system call to determine
whether the process is in a specific group (apart from the main group).

This was bad, because we don't want to use pids in system calls and
also because not everybody should know in which groups a specific
process is.

The new way is to pass the uid, gid and all other gids with the open
operation to devices and userspace filesystems. They can store that
and use it for permission checks in meta operations like mkdir and so
on, which are all based on open files anyway.
Another side-effect is that we can now, e.g., open a directory and
delegate that file descriptor to someone else to perform a meta
operation with that file descriptor, using the permissions of the
opener. I'm not sure yet where this is useful, but it fits nicely into
our combination of ACL and capabilities.

This also lead to quite some changes in the kernel, because it makes
much more sense now to pass this "user struct" around instead of the
process id.