kernel: removed pid argument from rd/wr/sd/recv.
This was only used for statistical purposes (input/output). Passing the
pid as an argument to all kinds of functions just for this purpose is
not a good idea. Instead, we do the statistics in the syscall functions
now. This changes the semantics of input/output stats slightly, because
it is now only the input/output directly caused by the process via
read/write/send/receive system calls. ELF loading, demand loading,
messages from e.g. delegate and so on, are no longer counted.